+++ gegelatilib/include/tpg/tpgExecutionEngine.h
@@ -147,8 +147,8 @@ namespace TPG {
          *         evaluation of the TPGGraph. The TPGAction resulting from the
          *         TPGGraph execution is at the end of the returned vector.
          */
-        virtual std::pair<std::vector<const TPG::TPGVertex*>, std::vector<size_t>> executeFromRoot(
-            const TPGVertex& root, const std::vector<size_t>& initActions, uint64_t nbEdgesActivated);
+        virtual std::pair<std::vector<const TPG::TPGVertex*>, std::vector<uint64_t>> executeFromRoot(
+            const TPGVertex& root, const std::vector<uint64_t>& initActions, uint64_t nbEdgesActivated);
     };
 }; // namespace TPG
