+++ gegelatilib/src/learn/learningAgent.cpp
@@ -153,7 +153,7 @@ std::shared_ptr<Learn::EvaluationResult> Learn::LearningAgent::evaluateJob(
         while (!le.isTerminal() &&
                nbActions < this->params.maxNbActionsPerEval) {
             // Get the actions
-            std::vector<size_t> actionsID 
+            std::vector<uint64_t> actionsID 
                 = tee.executeFromRoot(*root, le.getInitActions(), this->params.nbEdgesActivable).second;
 
             // Do it
